fix symbolic problem
